Allow the month field if the value of the field is zero
Generally sane systems will not output a value for month due to the difficulties of parsing it, however it is possible to come
across systems that will include a value of zero for the month.

Given that a zero-valued month has no effect on the output here, this change allows that case, so that you can still use this
library to parse an otherwise valid duration.